+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 * Add knife 'ssl check' and 'ssl fetch' commands for debugging SSL errors (CHEF-4711)
 * Cron resource accepts a weekday attribute as a symbol. (CHEF-4848)
 * Cron resource accepts special strings, e.g. @reboot (CHEF-2816)
+* Call WIN32OLE.ole_initialize before using WMI (CHEF-4888)
 
 ## Last Release: 11.10.0 (02/06/2014)